Work on migrating components
