+++ package.json
@@ -6,6 +6,9 @@
 		"@apollo/client": "^3.2.4",
 		"@auth0/auth0-react": "^1.1.0",
 		"@blockly/block-plus-minus": "^2.0.16",
+		"@emotion/cache": "^11.0.0",
+		"@emotion/react": "^11.1.1",
+		"@emotion/server": "^11.0.0",
 		"@material-ui/core": "^4.11.0",
 		"@material-ui/icons": "^4.9.1",
 		"@material-ui/lab": "^4.0.0-alpha.56",
@@ -36,6 +39,7 @@
 		"ml-svm": "^2.1.2",
 		"nanoid": "^3.1.16",
 		"next": "^10.0.2",
+		"next-cookies": "^2.0.3",
 		"plotly.js": "^1.55.2",
 		"react": "^17.0.1",
 		"react-blockly": "^5.3.0",
