+++ src/pages/_app.tsx
@@ -1,20 +1,33 @@
 import Head from 'next/head';
 import type { AppProps } from 'next/app';
-import React from 'react';
+import React, { useEffect } from 'react';
 import { DarkThemeProvider } from '../components/DarkThemeProvider';
 import { Auth0Provider } from '@auth0/auth0-react';
 import AppProviders from '../components/AppProviders';
+import createCache from '@emotion/cache';
+import { CacheProvider } from '@emotion/react';
+import { CssBaseline } from '@material-ui/core';
+
+export const cache = createCache({ key: 'css' });
 
 export default function MyApp({ Component, pageProps }: AppProps) {
-  const domain = process.env.REACT_APP_AUTH0_DOMAIN;
-  const clientId = process.env.REACT_APP_AUTH0_CLIENT_ID;
+  const domain = process.env.NEXT_PUBLIC_AUTH0_DOMAIN;
+  const clientId = process.env.NEXT_PUBLIC_AUTH0_CLIENT_ID;
 
   if(domain === undefined || clientId === undefined) {
-      throw new Error("REACT_APP_AUTH0_DOMAIN and/or REACT_APP_AUTH0_CLIENT_ID are/is undefined");
+      throw new Error("NEXT_PUBLIC_AUTH0_DOMAIN and/or NEXT_PUBLIC_AUTH0_CLIENT_ID are/is undefined");
   }
 
+  useEffect(() => {
+    // Remove the server-side injected CSS.
+    const jssStyles = document.querySelector('#jss-server-side');
+    if (jssStyles) {
+      jssStyles.parentElement?.removeChild(jssStyles);
+    }
+  }, []);
+
   return (
-    <>
+    <CacheProvider value={cache}>
       <Head>
         <link rel="icon" href="/favicon.ico" />
         <link
@@ -26,12 +39,14 @@ export default function MyApp({ Component, pageProps }: AppProps) {
           href="https://fonts.googleapis.com/icon?family=Material+Icons"
         />
       </Head>
+      <CssBaseline />
       <DarkThemeProvider>
+        { /* On server globalThis is defined but window isn't */}
         <Auth0Provider
           domain={ domain }
           clientId={ clientId }
-          redirectUri={ window.location.origin }
-          audience={ process.env.REACT_APP_AUTH0_AUDIENCE }
+          redirectUri={ globalThis.window?.location.origin ?? undefined }
+          audience={ process.env.NEXT_PUBLIC_AUTH0_AUDIENCE }
         >
           {/* Other providers that depend on hooks from providers in this component */}
           <AppProviders>
@@ -39,6 +54,6 @@ export default function MyApp({ Component, pageProps }: AppProps) {
           </AppProviders>
         </Auth0Provider>
       </DarkThemeProvider>
-    </>
+    </CacheProvider>
   );
 }
